requirements.txt
numpy
Pillow
tensorflow
requests==2.25.1
streamlit==0.76.0
click==8.0.4
google_api_python_client==1.12.8
protobuf==3.14.0